+++ app/src/main/java/com/example/android/popmovies/TrailersActivity.java
@@ -20,6 +20,12 @@ import android.view.MenuItem;
 import android.view.View;
 import android.widget.Toast;
 
+import com.example.android.popmovies.utilities.NetworkUtils;
+
+import org.json.JSONArray;
+import org.json.JSONObject;
+
+import java.net.URL;
 import java.util.ArrayList;
 import java.util.HashMap;
 
@@ -27,17 +33,17 @@ import java.util.HashMap;
  * Created by toda on 2017/12/18.
  */
 
-public class TrailersActivity extends AppCompatActivity implements TrailersAdapter.ListItemClickListener{
+public class TrailersActivity extends AppCompatActivity implements TrailersAdapter.ListItemClickListener {
 
     private static final String TAG = TrailersActivity.class.getSimpleName();
 
     private TrailersAdapter mAdapter;
     private RecyclerView mTrailersList;
-    private static final int NUM_LIST_ITEMS = 20;
     String mId;
-    private static ArrayList<HashMap<String, String>> trailersInfo;
+    private static ArrayList<HashMap<String, String>> trailersInfo = new ArrayList<>();
 
     private Toast mToast;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -57,24 +63,106 @@ public class TrailersActivity extends AppCompatActivity implements TrailersAdapt
 
         mTrailersList.setHasFixedSize(true);
 
-        mAdapter = new TrailersAdapter(NUM_LIST_ITEMS, this);
+        mAdapter = new TrailersAdapter(this);
         mTrailersList.setAdapter(mAdapter);
 
         loadMovieData();
     }
 
     @Override
-    public void onListItemClick(String  TrailersUrl) {
+    public void onListItemClick(String TrailersUrl) {
         Log.v(TAG, TrailersUrl);
     }
 
     private void loadMovieData() {
-        new MainActivityFragment.FetchLoadingTaskDetail().execute(getString(R.string.action_get_trailers),mId);
+        new FetchLoadingTaskDetail().execute(getString(R.string.action_get_trailers), mId);
     }
 
     private void showMovieDataView() {
         mTrailersList.setVisibility(View.VISIBLE);
     }
 
+
+    class FetchLoadingTaskDetail extends AsyncTask<String, Void, ArrayList<HashMap<String, String>>> {
+
+        private String API_KEY = "13a909a18259e3653d6885c366f6b369";
+
+        @Override
+        protected void onPreExecute() {
+            super.onPreExecute();
+        }
+
+        //        Start loading images on background
+        @Override
+        protected ArrayList<HashMap<String, String>> doInBackground(String... params) {
+
+            if (params.length == 0) {
+                return null;
+            }
+
+            String sortOrder = params[0];
+            String mId = params[1];
+
+            // clear the data of ArrayList if it's not empty
+            if (trailersInfo.size() > 0) {
+                trailersInfo.clear();
+            }
+
+            URL requestUrl = NetworkUtils.buildUrl_detail(API_KEY, sortOrder, mId);
+
+            try {
+                //Store JsonResponse
+                String jsonResponse = NetworkUtils
+                        .getResponseFromHttpUrl(requestUrl);
+
+                Log.v(TAG, "jsonResponse: " + jsonResponse);
+
+                //Find items in "Result"
+                JSONObject getMovieInfo = new JSONObject(jsonResponse);
+//                Log.v(TAG, getMovieInfo.getString("page"));
+                JSONArray resultDetail = getMovieInfo.getJSONArray("results");
+
+                Log.v(TAG, "JSON Result: " + resultDetail.toString());
+
+                // looping through All Contacts
+                for (int i = 0; i < resultDetail.length(); i++) {
+
+                    JSONObject r = resultDetail.getJSONObject(i);
+
+                    String id = r.getString("id");
+                    String key = r.getString("key");
+                    String name = r.getString("name");
+                    String type = r.getString("type");
+                    String trailersUrl = "https://www.youtube.com/watch?v=" + key;
+
+                    HashMap<String, String> movieInfo = new HashMap<>();
+                    // adding each child node to HashMap key => value
+
+                    movieInfo.put("id", id);
+                    movieInfo.put("name", name);
+                    movieInfo.put("trailersUrl", trailersUrl);
+
+                    // adding contact to movieInfo list
+                    trailersInfo.add(movieInfo);
+                }
+                Log.v(TAG, "MoviesInfo: " + trailersInfo.toString());
+                return trailersInfo;
+            } catch (Exception e) {
+                Log.e(TAG, "Json parsing error: " + e.getMessage());
+                e.printStackTrace();
+                return null;
+            }
+        }
+
+        // Post the loading data
+        @Override
+        protected void onPostExecute(ArrayList<HashMap<String, String>> hashMaps) {
+            super.onPostExecute(hashMaps);
+            if (hashMaps != null) {
+                Log.v(TAG, "onPost is executed" + hashMaps.toString());
+                mAdapter.setTrailersData(hashMaps);
+            }
+        }
+    }
 
 }
