+++ app/src/main/res/layout/trailers_list_item.xml
@@ -2,6 +2,7 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
+    xmlns:tools="http://schemas.android.com/tools"
     android:orientation="vertical">
 
     <!--This TextView holds the video data for one particular movie-->
@@ -24,6 +25,7 @@
             style="@style/TextAppearance.AppCompat.Large"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
+            tools:text="Trailer Name"
             android:padding="16dp"/>
 
     </LinearLayout>
